third_party/blink/web_tests/platform/generic/fast/backgrounds/repeat/parsing-background-repeat-expected.txt

This tests checks that all of the input values for background-repeat parse correctly.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS test("background-repeat: repeat-x;") is "repeat-x"
PASS test("background-repeat: repeat-y;") is "repeat-y"
PASS test("background-repeat: repeat;") is "repeat"
PASS test("background-repeat: no-repeat;") is "no-repeat"
PASS test("background-repeat: round;") is "round"
PASS test("background-repeat: space;") is "space"
PASS test("background-repeat: repeat repeat;") is "repeat"
PASS test("background-repeat: no-repeat space;") is "no-repeat space"
PASS test("background-repeat: round round;") is "round"
PASS test("background-repeat: space repeat;") is "space repeat"
PASS test("background: purple url(resources/gradient.gif) repeat-x top left") is "repeat-x"
PASS test("background: purple url(resources/gradient.gif) repeat-y 50% 50%") is "repeat-y"
PASS test("background: purple url(resources/gradient.gif) repeat center") is "repeat"
PASS test("background: purple url(resources/gradient.gif) no-repeat 12px") is "no-repeat"
PASS test("background: purple url(resources/gradient.gif) round left 50px") is "round"
PASS test("background: purple url(resources/gradient.gif) space 25px 25px") is "space"
PASS test("background-repeat: 45;") is ""
PASS test("background-repeat: coconut;") is ""
PASS successfullyParsed is true

TEST COMPLETE
